+++ src/dat-fs-provider/add-ftl-dat.ts
@@ -6,7 +6,11 @@ export async function addFtlDat() {
   if (!files) return;
   const datFile = files[0];
   workspace.updateWorkspaceFolders(
-      0,
+      // add the dat file as the last workspace folder
+      // this is because vscode uses the schema of the first folder
+      // to determine which file picker to use when adding
+      // folders to the workspace
+      workspace.workspaceFolders?.length ?? 0,
       0,
       {uri: datFile.with({scheme: FtlDatFs.scheme}), name: 'FTL Dat'}
   );