+++ CHANGELOG.md
@@ -91,3 +91,6 @@ All notable changes to the "ftlXml" extension will be documented in this file.
 - Add animation validator to check the frames used vs defined
 - fix bug where you couldn't add a folder to the workspace when an ftl.dat was opened already
 - allow hovering over an image path to see a preview of it
+
+## 0.4.3
+- Auto parse files when added, including png and audio files